requirements.txt
altair==5.4.1
attrs==24.2.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.2
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
entrypoints==0.4
Faker==28.4.1
favicon==0.7.0
fonttools==4.53.1
frozendict==2.4.4
gitdb==4.0.11
GitPython==3.1.43
htbuilder==0.6.2
html5lib==1.1
idna==3.8
isodate==0.6.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
msrest==0.7.1
multitasking==0.0.11
narwhals==1.6.0
numpy==2.1.0
oauthlib==3.2.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
peewee==3.17.6
pillow==10.4.0
platformdirs==4.3.1
plotly==5.24.1
prometheus_client==0.20.0
protobuf==5.28.0
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.9
pyparsing==3.1.4
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.38.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.7
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
validators==0.34.0
watchdog==4.0.2
webencodings==0.5.1
yfinance==0.2.43
